postgresql

+++ Programming/Django/address_django/address/migrations/0001_initial.py
@@ -16,7 +16,7 @@ class Migration(migrations.Migration):
             fields=[
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('address_line', models.CharField(max_length=40)),
-                ('postal_code', models.DecimalField(decimal_places=2, max_digits=5)),
+                ('postal_code', models.IntegerField()),
                 ('country', models.CharField(max_length=40)),
                 ('city', models.CharField(max_length=40)),
                 ('fax_number', models.CharField(max_length=13)),
